requirements.txt
asgiref==3.9.2
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pygame==2.6.1
PyJWT==2.10.1
pytest==8.3.2
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.3
tailwind==3.1.5b0
tzdata==2025.2
